cipher implementation + bcrypt cleanup
