+++ Sources/Crypto/Digest/DigestAlgorithm.swift
