+++ Sources/Crypto/Utilities/Deprecated.swift
@@ -1,0 +1,6 @@
+extension Digest {
+    @available(*, deprecated, renamed: "hash")
+    public func digest(_ data: LosslessDataConvertible) throws -> Data {
+        return try hash(data)
+    }
+}
