+++ Tests/CryptoTests/BCryptTests.swift
@@ -11,42 +11,39 @@ class BCryptTests: XCTestCase {
     ]
     
     func testVersion() throws {
-        let digest = try BCrypt.make(message: "foo", with: BCrypt.Salt(.two(.y), cost: 6, bytes: nil))
+        let digest = try BCrypt.hash("foo", cost: 6)
         XCTAssert(String(bytes: digest, encoding: .utf8)!.hasPrefix("$2y$06$"))
     }
     
     func testFail() throws {
-        let salt = try BCrypt.Salt(.two(.y), cost: 6, bytes: nil)
-        let digest = try BCrypt.make(message: "foo", with: salt)
-        let res = try BCrypt.verify(message: "bar", matches: digest)
+        let digest = try BCrypt.hash("foo", cost: 6)
+        let res = try BCrypt.verify("bar", matches: digest)
         XCTAssertEqual(res, false)
     }
     
     func testSanity() throws {
         let secret = "passwordpassword"
+        let res = try BCrypt.hash("foo", cost: 4, salt: secret)
+
+        let parser = try BCryptParser(serialized: res)
+        let parsedSalt = try parser.parseConfig()
         
-        let salt = try BCrypt.Salt(.two(.y), cost: 4, bytes: Data(secret.utf8))
-        let res = try BCrypt.make(message: "foo", with: salt)
-        
-        let parser = try BCrypt.Parser(res)
-        let parsedSalt = try parser.parseSalt()
-        
-        XCTAssertEqual(secret, String(bytes: parsedSalt.bytes, encoding: .utf8))
+        XCTAssertEqual(secret, String(bytes: parsedSalt.salt, encoding: .utf8))
     }
     
     func testInvalidSalt() throws {
         do {
-            _ = try BCrypt.Parser(Data("foo".utf8))
+            _ = try BCryptParser(serialized: Data("foo".utf8))
             XCTFail("Should have failed")
-        } catch let error as BCrypt.Error {
+        } catch let error as CryptoError {
             print(error)
         }
     }
     
     func testVerify() throws {
         for (desired, message) in tests {
-            let result = try BCrypt.verify(message: message, matches: desired)
-            XCTAssert(result, "Message '\(message)' did not create \(desired)")
+            let result = try BCrypt.verify(message, matches: desired)
+            XCTAssert(result, "\(message): did not match \(desired)")
         }
     }
 }
