+++ Tests/LinuxMain.swift
@@ -1,9 +1,16 @@
 import XCTest
 @testable import CryptoTests
+@testable import RandomTests
 
 XCTMain([
+	/// Crypto
+    testCase(BCryptTests.allTests),
+    testCase(CipherTests.allTests),
     testCase(MD5Tests.allTests),
+    testCase(RSATests.allTests),
     testCase(SHA1Tests.allTests),
     testCase(SHA2Tests.allTests),
-    testCase(RSATests.allTests),
+
+    /// Random
+    testCase(RandomTests.allTests),
 ])
